// apps/server/src/types.ts
export type ProductType = {
  id: string;
  category: string;
  coffee_type?: CoffeeType;
  image?: string;
  name: string;
  description?: string;
  price: number;
  sales: number;
};

export type CategoryType = {
  id: string;
  name: string;
  altName?: string;
  description: string;
};

export type CoffeeType = "Hot Coffee" | "Cold Coffee" | "Hot Chocolate";

export interface AddressType {
  id?: string;
  unit_number?: string;
  street_number?: string;
  street_name?: string;
  barangay: string;
  city: string;
  province?: string;
  region?: string;
  postal_code?: number;
  address_line2?: string;
  phone_number?: string;
  tel_number?: string;
  type: "home" | "work";
}

export interface DeliveryAddressType extends AddressType {
  name: string;
  user_uid: string;
}

export type UserType = {
  id: string;
  default_address: string;
  nickname: string;
  gender: string;
  favorites?: string[];
};

export type CartItemType = {
  index: number;
  product_snapshot: ProductType;
  product_id: string;
  quantity: number;
  size: Size;
  milk: Milk;
  syrup: Syrup;
  additives: Additive[];
  ice: Ice;
  name?: string;
};

export type OrderType = {
  id?: string;
  created_at: Timestamp;
  products: CartItemType[];
  total_price: number;
  user_uid: string;
  branch: BranchType;
  payment_status: PaymentStatusType;
  payment_at?: Timestamp;
  payment_option: PaymentOptionType;
  delivery_at?: Timestamp;
  delivery_status: DeliveryStatusType;
  delivery_option: DeliveryOptionType;
  delivery_address_id?: string;
  delivery_fee?: number;
};

export enum DeliveryStatusType {
  Pending = "Pending",
  Preparing = "Preparing",
  OnTheWay = "On the Way",
  Delivered = "Delivered",
  Cancelled = "Cancelled",
}

export type OrderProductType = {
  product_id: string;
  quantity: number;
  size?: Size;
};

export enum PaymentOptionType {
  OverTheCounter = "Over the Counter",
  GCash = "GCash",
  LazadaPay = "Lazada Wallet",
  ShopeePay = "Shopee Pay",
  PayMaya = "PayMaya",
  CoinsPH = "Coins.ph",
  GooglePay = "Google Pay",
  ApplePay = "Apple Pay",
  CreditCard = "Credit Card",
  DebitCard = "Debit Card",
}

export enum DeliveryOptionType {
  Pickup = "Pickup",
  Delivery = "Delivery",
}

export enum PaymentStatusType {
  Paid = "paid",
  Pending = "pending",
}

export type BranchType = {
  id?: string;
  name?: string;
  address: AddressType;
  main: boolean;
};
// ...
export enum Size {
  None = "None",
  Tall = "Tall",
  Grande = "Grande",
  Venti = "Venti",
}

export enum Milk {
  None = "None",
  Cow_s = "Cow's",
  Lactose_Free = "Lactose-Free",
  Skimmed = "Skimmed",
  Vegetable = "Vegetable",
}

export enum Syrup {
  None = "None",
  Amaretto = "Amaretto",
  Caramel = "Caramel",
  Hazelnut = "Hazelnut",
  Irish_Cream = "Irish Cream",
  Vanilla = "Vanilla",
}

export enum Additive {
  Ceylon_Cinnamon = "Ceylon Cinnamon",
  Grated_Chocolate = "Grated Chocolate",
  Liquid_Chocolate = "Liquid Chocolate",
  Marshmallow = "Marshmallow",
  Whipped_Cream = "Whipped Cream",
  Nutmeg = "Nutmeg",
  Ice_Cream = "Ice Cream",
}

export enum Ice {
  None = "None",
  Light = "Light",
  Normal = "Normal",
  Extra = "Extra",
}
// ...
export interface invoiceType {
  shipping: {
    name: string;
    address: string;
    city: string;
    country: string;
    postal_code: number;
  };
  items: {
    item: string;
    description: string;
    quantity: number;
    amount: number;
  }[];
  payment_method: PaymentOptionType;
  payment_status: PaymentStatusType;
  subtotal: number;
  paid: number;
  invoice_nr: string;
}